refactor(website): extract useEditorMode hook
